geom signif with lmm test. See geom_signif from ggsignif for more information

geom signif with lmm test. See geom_signif from ggsignif for more information

Usage

geom_signif_lmm(
  data_df,
  response,
  condition,
  latent_vars,
  comparisons,
  y_position = NULL,
  tip_length = 0.03,
  size = 0.5,
  step_increase = 0
)

Arguments

data_df

data frame containing data

response

name of column in data_df that corresponds to the response variable

condition

name of column in data_df that corresponds to condition of interest

latent_vars

vector containing columns in data_df that correspond to variables that describe the nested structure of the data e.g. Patient

comparisons

list of vectors of length two that define the comparisons to make

y_position

position of brackets (default=NULL)

tip_length

length of bracket tips (default = 0.03)

size

size of brackets (default = 0.5)

step_increase

distance between brackets (default = 0)

Value

geom signif

Examples